Add simple migration implementation and configurable selection
